+++ AoE2ScenarioParser/objects/data_objects/condition.py
@@ -14,10 +14,10 @@ class Condition(AoE2Object):
     _link_list = [
         RetrieverObjectLink("condition_type", "Triggers",
                             "trigger_data[__index__].condition_data[__index__].condition_type"),
-        RetrieverObjectLink("amount_or_quantity", "Triggers",
-                            "trigger_data[__index__].condition_data[__index__].amount_or_quantity"),
-        RetrieverObjectLink("resource_type_or_tribute_list", "Triggers",
-                            "trigger_data[__index__].condition_data[__index__].resource_type_or_tribute_list"),
+        RetrieverObjectLink("quantity", "Triggers",
+                            "trigger_data[__index__].condition_data[__index__].quantity"),
+        RetrieverObjectLink("attribute", "Triggers",
+                            "trigger_data[__index__].condition_data[__index__].attribute"),
         RetrieverObjectLink("unit_object", "Triggers",
                             "trigger_data[__index__].condition_data[__index__].unit_object"),
         RetrieverObjectLink("next_object", "Triggers",
@@ -52,8 +52,8 @@ class Condition(AoE2Object):
 
     def __init__(self,
                  condition_type: int = None,
-                 amount_or_quantity: int = None,
-                 resource_type_or_tribute_list: int = None,
+                 quantity: int = None,
+                 attribute: int = None,
                  unit_object: int = None,
                  next_object: int = None,
                  object_list: int = None,
@@ -75,8 +75,8 @@ class Condition(AoE2Object):
                  xs_function: str = None,
                  ):
         self.condition_type: int = condition_type
-        self.amount_or_quantity: int = amount_or_quantity
-        self.resource_type_or_tribute_list: int = resource_type_or_tribute_list
+        self.quantity: int = quantity
+        self.attribute: int = attribute
         self.unit_object: int = unit_object
         self.next_object: int = next_object
         self.object_list: int = object_list