+++ AoE2ScenarioParser/objects/data_objects/trigger.py
@@ -138,11 +138,11 @@ class Trigger(AoE2Object):
                     source_player=None, target_player=None, technology=None, string_id=None, display_time=None,
                     trigger_id=None, location_x=None, location_y=None, location_object_reference=None, area_1_x=None,
                     area_1_y=None, area_2_x=None, area_2_y=None, object_group=None, object_type=None,
-                    instruction_panel_position=None, attack_stance=None, time_unit=None, enabled_or_victory=None,
+                    instruction_panel_position=None, attack_stance=None, time_unit=None, enabled=None,
                     food=None, wood=None, stone=None, gold=None, item_id=None, flash_object=None,
                     force_research_technology=None, visibility_state=None, scroll=None, operation=None,
                     object_list_unit_id_2=None, button_location=None, ai_signal_value=None, object_attributes=None,
-                    from_variable=None, variable_or_timer=None, facet=None, play_sound=None, message=None,
+                    variable=None, timer=None, facet=None, play_sound=None, message=None,
                     player_color=None, sound_name=None, selected_object_ids=None) -> Effect:
         """Used to add new effect to trigger. Please use trigger.new_effect.<effect_name> instead"""
 
@@ -164,8 +164,8 @@ class Trigger(AoE2Object):
         self.effects.append(new_effect)
         return new_effect
 
-    def _add_condition(self, condition_type: ConditionId, amount_or_quantity=None,
-                       resource_type_or_tribute_list=None, unit_object=None, next_object=None, object_list=None,
+    def _add_condition(self, condition_type: ConditionId, quantity=None,
+                       attribute=None, unit_object=None, next_object=None, object_list=None,
                        source_player=None, technology=None, timer=None, area_1_x=None, area_1_y=None, area_2_x=None,
                        area_2_y=None, object_group=None, object_type=None, ai_signal=None, inverted=None, variable=None,
                        comparison=None, target_player=None, unit_ai_action=None, xs_function=None) -> Condition: